pyproject.toml
[project]
name = "mj_environment"
version = "1.1.0"
description = "A Python module for managing dynamic objects and full environments in MuJoCo"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Siddhartha Srinivasa", email = "siddhartha.srinivasa@gmail.com"},
]
maintainers = [
    {name = "Siddhartha Srinivasa", email = "siddhartha.srinivasa@gmail.com"},
]
requires-python = ">=3.9"
dependencies = [
    "mujoco>=2.1.0",
    "numpy",
    "pyyaml",
    "asset-manager @ git+https://github.com/personalrobotics/asset_manager.git"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[tool.uv.sources]
asset-manager = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/personalrobotics/mj_environment"
Repository = "https://github.com/personalrobotics/mj_environment"
Issues = "https://github.com/personalrobotics/mj_environment/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mj_environment"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"src/mj_environment/__init__.py" = ["F401"]
